+++ src/main/java/woowacourse/shoppingcart/application/ProductService.java
@@ -1,6 +1,5 @@
 package woowacourse.shoppingcart.application;
 
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.shoppingcart.dao.ProductDao;
