+++ src/main/java/woowacourse/shoppingcart/dao/ProductDaoImpl.java
@@ -1,0 +1,77 @@
+package woowacourse.shoppingcart.dao;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+import java.util.Objects;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.stereotype.Repository;
+import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.exception.InvalidProductException;
+
+@Repository
+public class ProductDaoImpl implements ProductDao {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    private final RowMapper<Product> productRowMapper = (resultSet, rowNumber) ->
+            new Product(
+                    resultSet.getLong("id"),
+                    resultSet.getString("name"),
+                    resultSet.getInt("price"),
+                    resultSet.getString("image_url")
+            );
+
+    public ProductDaoImpl(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Long save(final Product product) {
+        final String query = "INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)";
+        final GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            final PreparedStatement preparedStatement =
+                    connection.prepareStatement(query, new String[]{"id"});
+            preparedStatement.setString(1, product.getName());
+            preparedStatement.setInt(2, product.getPrice());
+            preparedStatement.setString(3, product.getImageUrl());
+            return preparedStatement;
+        }, keyHolder);
+
+        return Objects.requireNonNull(keyHolder.getKey()).longValue();
+    }
+
+    public Product findProductById(final Long productId) {
+        try {
+            final String query = "SELECT name, price, image_url FROM product WHERE id = ?";
+            return jdbcTemplate.queryForObject(query, (resultSet, rowNumber) ->
+                    new Product(
+                            productId,
+                            resultSet.getString("name"), resultSet.getInt("price"),
+                            resultSet.getString("image_url")
+                    ), productId
+            );
+        } catch (EmptyResultDataAccessException e) {
+            throw new InvalidProductException();
+        }
+    }
+
+    public List<Product> findProducts() {
+        final String query = "SELECT id, name, price, image_url FROM product";
+        return jdbcTemplate.query(query, productRowMapper);
+    }
+
+    public List<Product> findProducts(long size, long page) {
+        final String query = "SELECT id, name, price, image_url FROM product LIMIT ?,?";
+        long start = (page - 1) * size;
+        long end = size;
+        return jdbcTemplate.query(query, productRowMapper, start, end);
+    }
+
+    public void delete(final Long productId) {
+        final String query = "DELETE FROM product WHERE id = ?";
+        jdbcTemplate.update(query, productId);
+    }
+}
